Non-pi
this is an update to attempt to setup rpi on a non-pi

+++ nssetup.py
@@ -20,22 +20,27 @@ from source import core
 answer = ''
 option = 0
 menuopt = 0
+isrpi = ''
+errormes = ''
+needlamp = ''
 
 try:
 	print("Welcome to Network-Scout - An addition logging application for Artillery.")
 	print("If you are installing the client side, please download artillery first.\n")
 	print("OPTIONS: \n1. Install Network-Scout Server\n2. Install Network-Scout Client \n3. Uninstall Network Scout \n4. Exit")
-	menuopt = input("Please select one:    ")
 
-	if menuopt is 2 and os.path.isdir("/var/artillery/"):
+	###Menu used for installation of NSServer NSclient and Removal of Network Scout
+	menuopt = raw_input("Please select one:    ")
+
+	if menuopt is 1:
+		option = 1
+		pass
+	elif menuopt is 2 and os.path.isdir("/var/artillery/"):
 		option = 2
 		pass
 	elif menuopt is 2:
 		print "Please install artillery from github.com/TrustedSec."
 		sys.exit()
-	elif menuopt is 1:
-		option = 1
-		pass
 	elif menuopt is 3 and os.path.isdir("/var/networkscout/"):
 		option = 3
 		pass
@@ -47,31 +52,40 @@ try:
 	else:
 		print "Invalid option. Please try again."
 		sys.exit()
-	
+
 	if option == 2:
 		print("[*]********** Installing network-scout...")
 		core.kill_artillery()
-		os.mkdir("/var/networkscout")
-		subprocess.Popen("cp -r ns/* /var/networkscout/", shell=True).wait()
-		subprocess.Popen("sudo apt-get install python-rpi.gpio", shell=True).wait()
+		if os.mkdir("/var/networkscout"):
+			subprocess.Popen("cp -r ns/* /var/networkscout/", shell=True).wait()
+		elif os.isdir("/var/networkscout"):
+			errormes = raw_input("Network scout already exists. Do you want to continue?")
+			if errormes.lower() == 'y' or isrpi.lower() == 'yes':
+				subprocess.Popen("rm -rf /var/networkscout/", shell=True).wait()
+				subprocess.Popen("cp -r ns/* /var/networkscout/", shell=True).wait()
+			else:
+				pass
+		else:
+			pass
+
 
 		#modifying artillery
 		print("[*]**********Modding Artillery for NS logging...")
 		mod = open("ns/stuff/artilleryfunction", "r")
 		contents = mod.read()
-		
+
 		artillery = open("/var/artillery/src/core.py", "a")
 		artillery.write(contents)
 		artillery.close()
 		mod.close()
 
 		#Adding nslog to all parts of artillery
-		
+
 		core.modify_program("warn_the_good_guys","/var/artillery/src/harden.py","	nslog(warning)")
 		core.modify_program("warn_the_good_guys","/var/artillery/src/honeypot.py","		    nslog(subject)")
 		core.modify_program("warn_the_good_guys","/var/artillery/src/monitor.py","		    nslog(subject)")
 		core.modify_program("warn_the_good_guys","/var/artillery/src/ssh_monitor.py","				    nslog(subject)")
-			
+
 		print("[*]********** Creating Log Directory and File...")
 		# create the database directories if they aren't there
 		if not os.path.isdir("/var/artillery/log/"):
@@ -80,10 +94,10 @@ try:
 			filewrite = file("/var/artillery/log/logs.txt", "w")
         	filewrite.write(" ")
         	filewrite.close()
-		
+
 		# install to rc.local
 		print "[*]********** Adding Network-Scout into startup through init scripts..."
-		
+
 		if os.path.isdir("/etc/init.d"):
 			if not os.path.isfile("/etc/init.d/nsclient"):
 				fileopen = file("./ns/startup/startup_network_scout_client", "r")
@@ -95,44 +109,51 @@ try:
 				print "[*] Triggering update-rc.d on Network Scout to automatic start..."
 				subprocess.Popen("chmod +x /etc/init.d/nsclient", shell=True).wait()
 				subprocess.Popen("update-rc.d nsclient defaults", shell=True).wait()
-				
-		print "[*]********** Adding LCD controller into startup through init scripts..."
-		
-		if os.path.isdir("/etc/init.d"):
-			if not os.path.isfile("/etc/init.d/lcd_controller"):
-				fileopen = file("./ns/startup/lcd_controller", "r")
-				config = fileopen.read()
-				fileopen.close()
-				filewrite = file("/etc/init.d/lcd_controller", "w")
-				filewrite.write(config)
-				filewrite.close()
-				print "[*] Triggering update-rc.d on LCD Controller to automatic start..."
-				subprocess.Popen("chmod +x /etc/init.d/lcd_controller", shell=True).wait()
-				subprocess.Popen("update-rc.d lcd_controller defaults", shell=True).wait()
-				
-		print "[*]********** Adding Shutdown into startup through init scripts..."
-		
-		if os.path.isdir("/etc/init.d"):
-			if not os.path.isfile("/etc/init.d/shutdown_button"):
-				fileopen = file("./ns/startup/shutdown", "r")
-				config = fileopen.read()
-				fileopen.close()
-				filewrite = file("/etc/init.d/shutdown_button", "w")
-				filewrite.write(config)
-				filewrite.close()
-				print "[*] Triggering update-rc.d on Shutdown Button to automatic start..."
-				subprocess.Popen("chmod +x /etc/init.d/shutdown_button", shell=True).wait()
-				subprocess.Popen("update-rc.d shutdown_button defaults", shell=True).wait()
-		
+
+
+		isrpi = raw_input("Will you need to install the LCD and shutdown button modules?")
+
+		if isrpi.lower() == 'y' or isrpi.lower() == 'yes':
+
+			print "[*]********** Adding LCD controller into startup through init scripts..."
+			subprocess.Popen("sudo apt-get install python-rpi.gpio", shell=True).wait()
+			if os.path.isdir("/etc/init.d"):
+				if not os.path.isfile("/etc/init.d/lcd_controller"):
+					fileopen = file("./ns/startup/lcd_controller", "r")
+					config = fileopen.read()
+					fileopen.close()
+					filewrite = file("/etc/init.d/lcd_controller", "w")
+					filewrite.write(config)
+					filewrite.close()
+					print "[*] Triggering update-rc.d on LCD Controller to automatic start..."
+					subprocess.Popen("chmod +x /etc/init.d/lcd_controller", shell=True).wait()
+					subprocess.Popen("update-rc.d lcd_controller defaults", shell=True).wait()
+
+			print "[*]********** Adding Shutdown into startup through init scripts..."
+			if os.path.isdir("/etc/init.d"):
+				if not os.path.isfile("/etc/init.d/shutdown_button"):
+					fileopen = file("./ns/startup/shutdown", "r")
+					config = fileopen.read()
+					fileopen.close()
+					filewrite = file("/etc/init.d/shutdown_button", "w")
+					filewrite.write(config)
+					filewrite.close()
+					print "[*] Triggering update-rc.d on Shutdown Button to automatic start..."
+					subprocess.Popen("chmod +x /etc/init.d/shutdown_button", shell=True).wait()
+					subprocess.Popen("update-rc.d shutdown_button defaults", shell=True).wait()
+
+			subprocess.Popen("chmod 755 /var/networkscout/lcd_controller.py", shell=True).wait()
+			subprocess.Popen("chmod 755 /var/networkscout/shutdown.py", shell=True).wait()
+			subprocess.Popen("cp /var/networkscout/source/Adafruit_CharLCD.py /usr/lib/python2.7/", shell=True).wait()
+
+		else:
+			pass
 
 		print("[*]********** Adding access to scripts for init.d...")
-		subprocess.Popen("chmod 755 /var/networkscout/lcd_controller.py", shell=True).wait()
-		subprocess.Popen("chmod 755 /var/networkscout/shutdown.py", shell=True).wait()
 		subprocess.Popen("chmod 755 /var/networkscout/nsclient.py", shell=True).wait()
 		subprocess.Popen("rm /var/networkscout/nsserver.py", shell=True).wait()
-		subprocess.Popen("cp /var/networkscout/source/Adafruit_CharLCD.py /usr/lib/python2.7/", shell=True).wait()
-		
-		answer=raw_input("Do you wish to reboot your pi? [yes|no]    ")
+
+		answer=raw_input("Do you wish to reboot your computer? [yes|no] ")
 		if answer.lower() == 'y' or answer.lower() == 'yes':
 			subprocess.Popen("reboot", shell=True)
 		else:
@@ -142,13 +163,16 @@ try:
 		print "[*]**********  Network server is preparing to install..."
 		os.mkdir("/var/networkscout/")
 		subprocess.Popen("cp -r ns/* /var/networkscout/", shell=True).wait()
-		print "[*]********** Downloading LAMP Install Script..."
-		subprocess.Popen("sudo git clone https://github.com/LikeABoss-001/Raspberry-Pi-LAMP-Install-Script.git", shell=True).wait()
-		print "[*]********** INSTALLING LAMP..."
-		print"[!]This may take a few minutes. Feel free to get a coffee. [!]"
-		subprocess.Popen("sudo chmod +x /home/pi/Raspberry-Pi-LAMP-Install-Script/install.sh && /home/pi/Raspberry-Pi-LAMP-Install-Script/install.sh", shell=True).wait()
-		subprocess.Popen("rm -rf Raspberry-Pi-LAMP-Install-Script/",shell=True).wait()
-		subprocess.Popen("sudo apt-get install python-rpi.gpio", shell=True).wait()
+
+		needlamp = raw_input("Will you need to install LAMP onto this machine? If not, the program will assume the server is set up with MySQL.  ")
+		if needlamp.lower() == 'y' or needlamp.lower() == 'yes':
+		    print("[*]********** Downloading LAMP Install Script...")
+		    subprocess.Popen("sudo git clone https://github.com/LikeABoss-001/Raspberry-Pi-LAMP-Install-Script.git", shell=True).wait()
+		    print "[*]********** INSTALLING LAMP..."
+		    print"[!]This may take a few minutes. Feel free to get a coffee. [!]"
+		    subprocess.Popen("sudo chmod +x /home/pi/Raspberry-Pi-LAMP-Install-Script/install.sh && /home/pi/Raspberry-Pi-LAMP-Install-Script/install.sh", shell=True).wait()
+		    subprocess.Popen("rm -rf Raspberry-Pi-LAMP-Install-Script/",shell=True).wait()
+		    subprocess.Popen("sudo apt-get install python-rpi.gpio", shell=True).wait()
 
 		# install to rc.local
 		print "[*]********** Adding Network-Scout into startup through init scripts..."
@@ -163,59 +187,69 @@ try:
 				subprocess.Popen("chmod +x /etc/init.d/nsserver", shell=True).wait()
 				subprocess.Popen("update-rc.d nsserver defaults", shell=True).wait()
 
-		print "[*]********** Adding LCD controller into startup through init scripts..."
-		
-		if os.path.isdir("/etc/init.d"):
-			if not os.path.isfile("/etc/init.d/lcd_controller"):
-				fileopen = file("./ns/startup/lcd_controller", "r")
-				config = fileopen.read()
-				fileopen.close()
-				filewrite = file("/etc/init.d/lcd_controller", "w")
-				filewrite.write(config)
-				filewrite.close()
-				print "[*] Triggering update-rc.d on LCD Controller to automatic start..."
-				subprocess.Popen("chmod +x /etc/init.d/lcd_controller", shell=True).wait()
-				subprocess.Popen("update-rc.d lcd_controller defaults", shell=True).wait()
-				
-		# install to rc.local
-		print "[*]********** Adding Shutdown into startup through init scripts..."
-		
-		if os.path.isdir("/etc/init.d"):
-			if not os.path.isfile("/etc/init.d/shutdown_button"):
-				fileopen = file("./ns/startup/shutdown", "r")
-				config = fileopen.read()
-				fileopen.close()
-				filewrite = file("/etc/init.d/shutdown_button", "w")
-				filewrite.write(config)
-				filewrite.close()
-				print "[*] Triggering update-rc.d on Shutdown Button to automatic start..."
-				subprocess.Popen("chmod +x /etc/init.d/shutdown_button", shell=True).wait()
-				subprocess.Popen("update-rc.d shutdown_button defaults", shell=True).wait()
-		
-		print("[*]********** Adding access to scripts for init.d...")		
-		subprocess.Popen("chmod 755 /var/networkscout/lcd_controller.py", shell=True).wait()
-		subprocess.Popen("chmod 755 /var/networkscout/shutdown.py", shell=True).wait()
+
+
+		if isrpi.lower() == 'y' or isrpi.lower() == 'yes':
+			print "[*]********** Adding LCD controller into startup through init scripts..."
+			subprocess.Popen("sudo apt-get install python-rpi.gpio", shell=True).wait()
+			if os.path.isdir("/etc/init.d"):
+				if not os.path.isfile("/etc/init.d/lcd_controller"):
+					fileopen = file("./ns/startup/lcd_controller", "r")
+					config = fileopen.read()
+					fileopen.close()
+					filewrite = file("/etc/init.d/lcd_controller", "w")
+					filewrite.write(config)
+					filewrite.close()
+					print "[*] Triggering update-rc.d on LCD Controller to automatic start..."
+					subprocess.Popen("chmod +x /etc/init.d/lcd_controller", shell=True).wait()
+					subprocess.Popen("update-rc.d lcd_controller defaults", shell=True).wait()
+
+					print "[*]********** Adding Shutdown into startup through init scripts..."
+
+			if os.path.isdir("/etc/init.d"):
+				if not os.path.isfile("/etc/init.d/shutdown_button"):
+					fileopen = file("./ns/startup/shutdown", "r")
+					config = fileopen.read()
+					fileopen.close()
+					filewrite = file("/etc/init.d/shutdown_button", "w")
+					filewrite.write(config)
+					filewrite.close()
+					print "[*] Triggering update-rc.d on Shutdown Button to automatic start..."
+					subprocess.Popen("chmod +x /etc/init.d/shutdown_button", shell=True).wait()
+					subprocess.Popen("update-rc.d shutdown_button defaults", shell=True).wait()
+			subprocess.Popen("chmod 755 /var/networkscout/lcd_controller.py", shell=True).wait()
+			subprocess.Popen("chmod 755 /var/networkscout/shutdown.py", shell=True).wait()
+			subprocess.Popen("chmod 755 /var/networkscout/lcd_controller.py", shell=True).wait()
+			subprocess.Popen("chmod 755 /var/networkscout/shutdown.py", shell=True).wait()
+
+			#moving Adafruit into python library
+			print("*********************** Putting the Pieces Together ********************")
+			subprocess.Popen("cp /var/networkscout/source/Adafruit_CharLCD.py /usr/lib/python2.7/", shell=True).wait()
+
+		else:
+			pass
+
+		print("[*]********** Adding access to scripts for init.d...")
+
 		subprocess.Popen("chmod 755 /var/networkscout/nsserver.py", shell=True).wait()
-				
-		#moving Adafruit into python library
-		print("*********************** Putting the Pieces Together ********************")
-		subprocess.Popen("cp /var/networkscout/source/Adafruit_CharLCD.py /usr/lib/python2.7/", shell=True).wait()
-		subprocess.Popen("mv /var/networkscout/website/* /var/www/", shell=True).wait()
-		subprocess.Popen("apt-get install python-mysqldb", shell=True).wait()
 		subprocess.Popen("rm /var/networkscout/nsclient.py", shell=True).wait()
-				
+		if os.path.isdir("/var/www/"):
+			subprocess.Popen("mv /var/networkscout/website/* /var/www/", shell=True).wait()
+		elif:
+			subprocess.Popen("mv /var/networkscout/website/* /var/www/html", shell=True).wait()
+
 		print("************************** Creating Database for Logs ***********************")
 		subprocess.Popen("python /var/networkscout/stuff/mysqltablecreator.py", shell=True).wait()
-		
+
 		serverip = core.ipgrab()
 		print("Website created at "+serverip+"/scoutwebsite.php \n")
-		
+
 		answer=raw_input("Do you wish to reboot your pi? [yes|no]    ")
 		if answer.lower() == 'y' or answer.lower() == 'yes':
 			subprocess.Popen("reboot", shell=True)
 		else:
 			pass
-				
+
 	elif option == 3:
 		answer = raw_input("Do you want to uninstall network-scout: [ yes | no }    ")
 		if answer.lower() == "y" or answer.lower() == "yes":
@@ -231,12 +265,12 @@ try:
 				subprocess.Popen("python /var/networkscout/stuff/mysqluninstaller.py", shell=True)
 				subprocess.Popen("rm /var/www/*", shell=True)
 				subprocess.Popen("apt-get purge `dpkg -l | awk -F ' ' ' /php|mysql|otherpackages/ { print $2 } '`", shell=True)
-				
+
 			print "[*] Network-Scout has been uninstalled. Manually kill the process if it is still running."
-				
+
 	else:
 		print "There was an issue installing Network-Scout."
-		
+
 except Exception, e:
 	print("There was an issue installing network-scout") + format(e)
-	sys.exit() 
+	sys.exit()
